fix: report delete
